+++ src/admin/listeners.js
@@ -39,9 +39,10 @@ export const onInteractionCreate = (interaction) => {
     return;
   }
 
+  const client = interaction.client; //get client
+
   if (interaction.isContextMenu()) {
     //context commands
-    const client = interaction.client; //get client
     const contextCommands = client.contextCommands; //get commands
 
     const foundCommand = contextCommands.find(
@@ -52,7 +53,21 @@ export const onInteractionCreate = (interaction) => {
     return;
   }
 
-  if (interaction.isCommand()) {
+  const slashCommands = client.slashCommands;
+
+  if (interaction.isAutocomplete()) {
+    //interaction with autocomplete activated
+    const autoCompleteCommands = slashCommands.filter(
+      (cmd) => cmd.autocomplete
+    ); //get commands with autocomplete action
+    const foundCommand = autoCompleteCommands
+      ? autoCompleteCommands.find(
+          (cmd) => cmd.command.name === interaction.commandName
+        )
+      : null; //find command that fired onInteractionCreate
+    if (foundCommand) foundCommand.autocomplete(interaction);
+    else interaction.respond([]); //if not found, return no choices
+  } else if (interaction.isCommand()) {
     //slash commands
     const client = interaction.client; //get client
     const slashCommands = client.slashCommands; //get commands
@@ -451,7 +466,8 @@ export const onMessageDelete = async (message) => {
         embed.addFields({
           name: messageDel.text,
           value: str,
-        }); //name's different from others
+        });
+      //name's different from others
       else embed.addFields({ name: messageDel.textAgain, value: str });
     });
   } else embed.addFields({ name: messageDel.text, value: content });
@@ -638,7 +654,8 @@ export const onMessageUpdate = async (oldMessage, newMessage) => {
             embed.addFields({
               name: messageU.contentOld,
               value: str,
-            }); //name's different from others
+            });
+          //name's different from others
           else embed.addFields({ name: messageU.contentOldAgain, value: str });
         });
       } else embed.addFields({ name: messageU.contentOld, value: oldContent });
@@ -653,7 +670,8 @@ export const onMessageUpdate = async (oldMessage, newMessage) => {
             embed.addFields({
               name: messageU.contentNew,
               value: str,
-            }); //name's different from others
+            });
+          //name's different from others
           else embed.addFields({ name: messageU.contentNewAgain, value: str });
         });
       } else embed.addFields({ name: messageU.contentNew, value: newContent });
@@ -785,7 +803,7 @@ export const onGuildMemberRemove = async (memberKick) => {
   const logCreationDate = kickLog ? dayjs(kickLog.createdAt) : null;
   const diff =
     logCreationDate !== null ? dayjs().diff(logCreationDate, "s") : null;
-  console.log("memberKick diff", diff)
+  console.log("memberKick diff", diff);
 
   //get user roles
   const roles = memberKick.roles.cache;